fix: remove unnecessary dependency and add gradle configuration file
References: https://outsystemsrd.atlassian.net/browse/RMET-539

The dependency removed pointed to a deprecated repository that provided google services functionality. The cordova-android version on this dependency was breaking the MABS build. Added build.gradle file to include google services. This is for MABS 6, for MABS 7 a Preference in the plugin.xml would be enough